Add cache class for local storage abstraction
This borrows from Laravel but for frontend requests.

This should only be used for non-sensitive data!

```
const articles = new Cache('key', async () => await fetch());
this.articles = await articles.get();
```

This allows for saving the result of requests to a local storage value
that will be used the next time the key is checked.

You can also check the size, if it's empty, and remove the value.